requirements.txt
azure-functions>=1.18.0
numpy>=1.26.0
